feat: organize server queries into separate files and enhance gamification profile data retrieval
