+++ src/server/queries/images.ts
@@ -1,0 +1,19 @@
+import "server-only";
+import { db } from "../db";
+
+export async function getAllImages() {
+  const images = await db.query.ImagesTable.findMany({
+    orderBy: (model, { desc }) => desc(model.id),
+  });
+  return images;
+}
+
+export async function getImage(id: number) {
+  const image = await db.query.ImagesTable.findFirst({
+    where: (model, { eq }) => eq(model.id, id),
+  });
+
+  if (!image) throw new Error("Image not found");
+
+  return image;
+}
